moved file App.js into components folder
