+++ src/components/App/index.js
@@ -2,15 +2,15 @@ import React, { Fragment } from 'react'
 import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
 
-import './config/reactotron'
+import '../../config/reactotron'
 
-import store from './store'
+import store from '../../store'
 
-import './styles/global'
+import '../../styles/global'
 
-import Routes from './routes'
+import Routes from '../../routes'
 
-import Header from './components/Header'
+import Header from '../../components/Header'
 
 const App = () => (
   <Provider store={store}>
